more on the previous commit

+++ hw6 updated/backend/static/script.js
@@ -6,9 +6,9 @@ var obj;
 var numValidCards = 0
 
 function valueCheck() {
-    let lowPrice = parseInt(document.getElementById('lowPrice').value)
-    let highPrice = parseInt(document.getElementById('highPrice').value)
-    if (lowPrice < 0) {
+    let lowPrice = parseFloat(document.getElementById('lowPrice').value)
+    let highPrice = parseFloat(document.getElementById('highPrice').value)
+    if (lowPrice < 0.0) {
         alert("Price Range values cannot be negative! Please try a value greater than or equal to 0.0")
         return false
     }
